+++ ebean-querybean/pom.xml
@@ -103,11 +103,11 @@
       <plugin>
         <groupId>io.repaint.maven</groupId>
         <artifactId>tiles-maven-plugin</artifactId>
-        <version>2.24</version>
+        <version>2.34</version>
         <extensions>true</extensions>
         <configuration>
           <tiles>
-            <tile>io.ebean.tile:enhancement:13.13.1</tile>
+            <tile>io.ebean.tile:enhancement:13.13.2</tile>
           </tiles>
         </configuration>
       </plugin>
