+++ querybean-generator/src/main/java/io/ebean/querybean/generator/SimpleQueryBeanWriter.java
@@ -117,7 +117,6 @@ class SimpleQueryBeanWriter {
    * Write the type query bean (root bean).
    */
   void writeRootBean() throws IOException {
-
     gatherPropertyDetails();
     if (isEmbeddable()) {
       processingContext.addEntity(beanFullName, dbName);
@@ -170,7 +169,11 @@ class SimpleQueryBeanWriter {
     importTypes.remove(Constants.DATABASE);
     importTypes.remove(Constants.FETCHGROUP);
     importTypes.remove(Constants.QUERY);
-    importTypes.add(Constants.TQASSOCBEAN);
+    if (embeddable) {
+      importTypes.add(Constants.TQASSOC);
+    } else {
+      importTypes.add(Constants.TQASSOCBEAN);
+    }
     if (isEntity()) {
       importTypes.add(Constants.TQPROPERTY);
       importTypes.add(origDestPackage + ".Q" + origShortName);
@@ -288,10 +291,7 @@ class SimpleQueryBeanWriter {
 
   private void writeAssocBeanFetch() {
     if (isEntity()) {
-      writeAssocBeanFetch("", "Eagerly fetch this association loading the specified properties.");
-      writeAssocBeanFetch("Query", "Eagerly fetch this association using a 'query join' loading the specified properties.");
-      writeAssocBeanFetch("Cache", "Eagerly fetch this association using L2 cache.");
-      writeAssocBeanFetch("Lazy", "Use lazy loading for this association loading the specified properties.");
+      // inherit the fetch methods
       if (implementsInterface != null) {
         writeAssocBeanExpression(false, "eq", "Is equal to by ID property.");
         writeAssocBeanExpression(true, "eqIfPresent", "Is equal to by ID property if the value is not null, if null no expression is added.");
@@ -367,8 +367,11 @@ class SimpleQueryBeanWriter {
       writer.append(" */").eol();
       writer.append(Constants.AT_GENERATED).eol();
       writer.append(Constants.AT_TYPEQUERYBEAN).eol();
-      writer.append("public class Q%s<R> extends TQAssocBean<%s,R> {", shortName, shortInnerName).eol();
-
+      if (embeddable) {
+        writer.append("public class Q%s<R> extends TQAssoc<%s,R> {", shortName, shortInnerName).eol();
+      } else {
+        writer.append("public class Q%s<R> extends TQAssocBean<%s,R,Q%s> {", shortName, shortInnerName, origShortName).eol();
+      }
     } else {
       writer.append("/**").eol();
       writer.append(" * Query bean for %s.", shortName).eol();
@@ -420,7 +423,6 @@ class SimpleQueryBeanWriter {
    * Write all the imports.
    */
   private void writeImports() {
-
     for (String importType : importTypes) {
       writer.append("import %s;", importType).eol();
     }
